Move interfaces into intercases.go an initialise in auth.go
