+++ oauth.go
@@ -14,37 +14,6 @@ var (
 	ErrNotImplemented   = errors.New("oauth feature not implemented yet")
 )
 
-// OAuthProvider represents configuration for an OAuth provider
-type OAuthProvider struct {
-	ClientID     string   `json:"client_id"`
-	ClientSecret string   `json:"client_secret"`
-	RedirectURL  string   `json:"redirect_url"`
-	Scopes       []string `json:"scopes"`
-}
-
-// OAuthConfig represents the overall OAuth configuration
-type OAuthConfig struct {
-	Providers    map[string]OAuthProvider `json:"providers"`
-	SessionStore sessions.Store           `json:"-"`
-	BaseURL      string                   `json:"base_url"`
-	SuccessURL   string                   `json:"success_url"`
-	FailureURL   string                   `json:"failure_url"`
-}
-
-// OAuthService defines the interface for OAuth operations
-type OAuthService interface {
-	// Provider management
-	RegisterProvider(name string, provider goth.Provider)
-	GetProvider(name string) (goth.Provider, error)
-	
-	// OAuth flow handlers
-	BeginAuthHandler() gin.HandlerFunc
-	CompleteAuthHandler() gin.HandlerFunc
-	
-	// User mapping
-	MapGothUserToUserInfo(gothUser goth.User) (UserInfo, error)
-}
-
 // oauthService implements the OAuthService interface
 type oauthService struct {
 	providers    map[string]goth.Provider
